Reorganized DataLimits files
