+++ src/EStimLibrary/Core/SpatialModel/IBodyModel.cs
@@ -1,4 +1,4 @@
-﻿using EStimLibrary.Core;
+﻿using EStimLibrary.Core.Data;
 
 
 namespace EStimLibrary.Core.SpatialModel;
