+++ src/EStimLibrary/EStimLibrary.csproj
@@ -25,6 +25,8 @@
     <None Remove="Extensions\Stimulation\Patterns\" />
     <None Remove="Extensions\Stimulation\Trains\" />
     <None Remove="Extensions\SpatialModel\StringHierarchy\" />
+    <None Remove="Core\Data\" />
+    <None Remove="Extensions\Data\" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Core\" />
@@ -44,6 +46,8 @@
     <Folder Include="Extensions\Stimulation\Patterns\" />
     <Folder Include="Extensions\Stimulation\Trains\" />
     <Folder Include="Extensions\SpatialModel\StringHierarchy\" />
+    <Folder Include="Core\Data\" />
+    <Folder Include="Extensions\Data\" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="System.IO.Ports" Version="7.0.0" />
