+++ src/EStimLibrary/Extensions/SpatialModel/StringHierarchy/StringHierarchyBodyModel.cs
@@ -1,4 +1,5 @@
 ﻿using EStimLibrary.Core;
+using EStimLibrary.Core.Data;
 using EStimLibrary.Core.SpatialModel;
